Working solution with lots of TODOs
