+++ SpatialGDK/Source/SpatialGDK/Public/SpatialCommonTypes.h
@@ -24,3 +24,12 @@ using FChannelObjectPair = TPair<TWeakObjectPtr<class USpatialActorChannel>, TWe
 struct FObjectReferences;
 using FObjectReferencesMap = TMap<int32, FObjectReferences>;
 using FReliableRPCMap = TMap<Worker_RequestId_Key, TSharedRef<struct FReliableRPCForRetry>>;
+
+// TODO(Alex): move it elsewhere?
+enum class ESpatialConnectionType
+{
+	Receptionist,
+	LegacyLocator,
+	Locator
+};
+
